refac: update project structure
