+++ package.json
@@ -38,6 +38,7 @@
     "react-native-screens": "^3.17.0",
     "react-native-splash-screen": "^3.3.0",
     "react-native-svg": "^13.2.0",
+    "react-native-toast-message": "^2.1.5",
     "zustand": "^4.1.1"
   },
   "devDependencies": {
