+++ src/navigation/main/BottomTab/index.tsx
@@ -2,12 +2,11 @@ import React from 'react';
 import {createBottomTabNavigator} from '@react-navigation/bottom-tabs';
 import screens from '../../../screens';
 import TabBar from './TabBar';
-import {ProjectParam} from '..';
 
 export type BottomStackParamList = {
   Calendar: undefined;
   MyPage: undefined;
-  Project: ProjectParam;
+  Project: undefined;
   Feedback: undefined;
 };
 
@@ -20,11 +19,7 @@ const BottomTabNavigator = () => {
       }}
       tabBar={props => <TabBar {...props} />}
       initialRouteName="Project">
-      <BottomTab.Screen
-        name="Project"
-        component={screens.Project}
-        initialParams={{type: 'main'}}
-      />
+      <BottomTab.Screen name="Project" component={screens.Project} />
       <BottomTab.Screen name="Calendar" component={screens.Calendar} />
       <BottomTab.Screen name="Feedback" component={screens.Feedback} />
       <BottomTab.Screen name="MyPage" component={screens.MyPage} />
