--- src/screens/Home/index.tsx
@@ -1,12 +1,0 @@
-import {View, Text} from 'react-native';
-import React from 'react';
-
-const Home = () => {
-  return (
-    <View>
-      <Text>Home</Text>
-    </View>
-  );
-};
-
-export default Home;
